Major Update
o Debug of RTC power up
o LEDs now indicate hardware failures
o Green LED now indicates if antenna is attached

+++ Software/AVR128DA48/FlexFox80/include/adc.h
@@ -39,6 +39,6 @@ void ADC0_setADCChannel(ADC_Active_Channel_t chan);
 void ADC0_startConversion(void);
 bool ADC0_conversionDone(void);
 int ADC0_read(void);
-float temperatureC(void);
+int16_t temperatureC(void);
 
 #endif /* ADC_H_ */
